37.AMQP Priority Exchange The idea behind our async and async_priority_high transports was that we can send some messages to async_priority_high and others to async, with the goal that those messages would end up in different "buckets"... or, more technically, in different "queues". Then we can instruct our worker to first read all messages from whatever queue async_priority_high is bound to before reading messages from whatever queue the async transport is bound to.
The queue_name Option in Doctrine

This did work before with Doctrine, thanks to this queue_name: high option. The default value for this option is... default. As a reminder, I'll quickly log into my database:

mysql -u root messenger_tutorial
And see what that table looked like:

DESCRIBE messenger_messages;
Yep, the queue_name column was the key to making this work. Messages that were sent to async_priority_high had a queue_name set to high, and those sent to the async transport had a value of default. So even though we only had one database table, it functioned like two queues: when we consumed the async_priority_high transport, it queried for all messages WHERE queue_name="high".

The problem is that this queue_name option is specific to the doctrine transport, and it has absolutely no effect when using AMQP.

Routing Messages to... a Queue?

But... on a high-level... our goal is the same: we need two queues. We need the async_priority_high transport to send messages to one queue and the async transport to send messages to a different queue.

But with AMQP... you don't send a message directly to a queue... you send it to an exchange... and then it's the exchange's responsibility to look at its internal rules and figure out which queue, or queues, that message should actually go to.

This means that to get a message to a queue, we need to tweak things on the exchange level. And there are two different ways to do this. First, we could continue to have a single exchange and then add some internal rules - called bindings - to teach the exchange that some messages should go to one queue and other messages should go to another queue. I'm going to show you how to do this a bit later.

The second option isn't quite as cool, but it's a bit simpler. By default, when Messenger creates an exchange, it creates it as a fanout type. That means that when a message is sent to this exchange, it's routed to every queue that's bound to it. So if we added a second binding to a second queue - maybe messages_high_priority - then every message that's sent to this exchange would be routed to both queues. It would be duplicated! That's... not what we want.

Instead, we're going to create two fanout exchanges, and each exchange will route all of its messages to a separate queue. We'll have two exchanges and two queues.