38.Delaying in AMQP: Dead Letter Exchange When we started working with AMQP, I told you to go into ImagePostController and remove the DelayStamp. This stamp is a way to tell the transport system to wait at least 500 milliseconds before allowing a worker to receive the message. Let's change this to 10 seconds - so 10000 milliseconds.
Now, move over to your terminal and make sure that your worker is not running.

Ok, let's see what happens! Right now both queues are empty. I'll upload 3 photos... then... quick, quick, quick! Go look at the queues. Suddenly, poof! A new queue appeared... with a strange name: delay_messages_high_priority__10000. And it has - dun, dun, dun! - three messages in it.

Let's look inside. Interesting, the messages were delivered here, instead of the normal queue. But then... they disappeared? The graph shows how the messages sitting in this queue went from 3 to 0. But... how? Our worker isn't even running!

Woh! This page just 404'ed! The queue is gone! Something is attacking our queues!

Head back to the queue list. Yea, that weird "delay" queue is gone... oh, but now the three messages are somehow in messages_high. What the heck just happened?

Well first, to prove that the whole system still works... regardless of what craziness just occurred... let's run our worker and consume from both the async_priority_high and async transports:

php bin/console -vv async_priority_high async
It consumes them and... when we move over, go to the homepage and refresh, yep! Ponka was added to those images.

The Delay Exchange

Ok, let's figure out how this worked. I mean, on the one hand, it's not important: if we had been running our worker the entire time, you would have seen that those messages were in fact delayed by 10 seconds. How you delay messages in RabbitMQ is kinda crazy... but if you don't care about the details, Messenger just takes care of it for you.

But I do want to see how this works... in part because it'll be a great chance to see how some of the more advanced features of AMQP work.

Click on "Exchanges". Surprise! There's a new exchange called delays. And instead of being a fanout type like our other two exchanges, this is a direct exchange. We'll talk about what that that means soon.

But the first thing to know is that when Messenger sees that a message should be delayed, it sends it to this exchange instead of sending it to the normal, "correct" exchange. At this moment, the delays exchange has no bindings... but that will change when we send a delayed message.