39.Exchange Routing and Binding Keys Let's change this delay back to one second... so we're not waiting all day for our photos to be processed.
Simple Setup: 1 Fanout Exchange per Queue

In messenger.yaml, the messages sent to each transport - async and async_priority_high - need to ultimately be delivered into two different queues so that we can consume them independently. And... we've accomplished that!

But there are two different ways that we could have done this. First, remember that in AMQP, messages are sent to an exchange, not a queue. Right now, when a message is routed to the async transport, Messenger sends that to an exchange called messages. You don't see that config here only because messages is the default exchange name in Messenger.

When a message is routed to the async_priority_high transport, Messenger sends that to an exchange called messages_high_priority. Each transport always sends to exactly one exchange.

Then, each exchange routes every message to a single queue, like the messages exchange sends to a messages queue... and messages_high_priority sends to a messages_high queue. There is not a routing key on the binding: Messenger binds each exchange to one queue... but with no routing key. That's how a "fanout" exchange works: it doesn't care about routing keys... it just sends each message to every queue bound to it.